+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Don't reset browse query when on Browse route and click Browse segment. Fixes UIIN-2434.
 * Prevent double-escaping of query when Browsing. Fixes UIIN-2435.
 * When adding items, cursor is in the barcode field as default. Refs UIIN-2205.
+* Quick export from instance detail view. Refs UIIN-2430.
 
 ## [9.4.5](https://github.com/folio-org/ui-inventory/tree/v9.4.5) (2023-04-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.4...v9.4.5)